Move all page components into folders to contain future unit tests.
