+++ src/pages/AuthPage/AuthPage.jsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
-import { signUp, signIn } from "../services/auth";
-import { validateAuthForm } from "../utils/validation";
-import PasswordInput from "../components/PasswordInput";
+import { signUp, signIn } from "../../services/auth";
+import { validateAuthForm } from "../../utils/validation";
+import PasswordInput from "../../components/PasswordInput/PasswordInput";
 
 const AuthPage = ({ setLoginMessage }) => {
   // Form input states
